pyproject.toml
[project]
name = "sotopia"
version = "0.1.0-rc.5"
description = "A platform for simulating and evaluating social interaction."
authors = [
    { name = "Hao Zhu", email = "prokilchu@gmail.com" },
    { name = "Xuhui Zhou", email = "xuhuiz@cs.cmu.edu" }
]
requires-python = ">=3.10, <3.13"
license = { text = "MIT License" }
readme = "README.md"

dependencies = [
    "lxml>=4.9.3,<6.0.0",
    "openai>=1.11.0,<2.0.0",
    "langchain>=0.2.5,<0.4.0",
    "rich>=13.6.0,<14.0.0",
    "PettingZoo==1.24.3",
    "redis-om>=0.3.0,<0.4.0",
    "gin-config>=0.5.0,<0.6.0",
    "absl-py>=2.0.0,<3.0.0",
    "together>=0.2.4,<0.3.0",
    "pydantic>=2.5.0,<3.0.0",
    "beartype>=0.12.0,<0.15.0",
    "langchain-openai>=0.1.8,<0.3.0",
    "litellm>=1.23.12,<1.50.0",
    "hiredis>=3.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
anthropic = ["anthropic"]
groq = ["groq"]
cohere = ["cohere"]
google-generativeai = ["google-generativeai"]
examples = ["transformers", "datasets", "scipy", "torch", "pandas"]
chat = ["fastapi", "uvicorn[standard]"]
test = ["pytest", "pytest-cov", "pytest-asyncio"]

[tool.uv]
dev-dependencies = [
    "pre-commit",
    "nbmake",
    "types-setuptools",
    "types-requests",
    "types-tqdm",
    "types-protobuf==4.24.0.4",
    "lxml-stubs",
    "pandas-stubs",
    "ruff",
    "mypy"
]

[tool.mypy]
mypy_path = "stubs"
strict = true
plugins = [
    "pydantic.mypy",
    "pydantic.v1.mypy"
]

[[tool.mypy.overrides]]
module = "transformers.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[project.scripts]
sotopia = "sotopia.cli:app"
